requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
openpyxl==3.1.5
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
aiofiles==24.1.0
jinja2==3.1.4
